+++ src/encoded/static/components/item-pages/components/file-overview/FileViewTabs.js
@@ -52,7 +52,7 @@ const AssociatedFilesTab = (props) => {
             ) : (
                 <div className="no-results">
                     <div className="no-results-content">
-                        <i class="icon icon-folder-open fas"></i>
+                        <i className="icon icon-folder-open fas"></i>
                         <h3 className="header">No Associated Files Found</h3>
                         <span className="subheader">
                             There are no DAC generated or externally generated
@@ -70,7 +70,7 @@ const AnalysisInformationTab = ({ context }) => {
     return (
         <div className="no-results">
             <div className="no-results-content">
-                <i class="icon icon-network-wired fas"></i>
+                <i className="icon icon-network-wired fas"></i>
                 <h3 className="header">Analysis Information Coming Soon</h3>
                 <span className="subheader">
                     Check back for updates on Analysis Information development
@@ -86,7 +86,7 @@ const QCOverviewTab = ({ context }) => {
     return (
         <div className="no-results">
             <div className="no-results-content">
-                <i class="icon icon-chart-area fas"></i>
+                <i className="icon icon-chart-area fas"></i>
                 <h3 className="header">QC Overview Coming Soon</h3>
                 <span className="subheader">
                     Check back for updates on QC Overview development with